Divide actions with batch actions
