+++ src/dashboard/src/hooks/useActions.ts
@@ -44,21 +44,6 @@ const useActions = (clusterId: string) => {
     })
   }, [clusterId]);
 
-  const batchUpdateStatus = useCallback((jobIds: string[], statusValue: string) => {
-    const url = `/api/clusters/${clusterId}/jobs/status`;
-    const status = Object.create(null);
-    for (const jobId of jobIds) {
-      status[jobId] = statusValue
-    }
-    return fetch(url, {
-      method: 'POST',
-      headers: {
-        'Content-Type': 'application/json'
-      },
-      body: JSON.stringify({ status })
-    })
-  }, [clusterId]);
-
   const onSupport = useCallback((event: any, job: any) => {
     const subject = `[DLTS Job][${clusterId}][${job['vcName']}]: <Issue Title by User>`;
     const body = `
@@ -139,78 +124,6 @@ ${givenName} ${familyName}
     });
   }, [confirm, enqueueSnackbar, updateStatus]);
 
-  const onBatchApprove = useCallback((event: any, jobs: any[]) => {
-    const title = `${jobs.length} jobs`;
-    return confirm(`Approve ${title}?`).then((answer) => {
-      if (answer === false) return;
-
-      const jobIds = jobs.map((job) => job['jobId'])
-
-      enqueueSnackbar(`${title} are being approved.`);
-      return batchUpdateStatus(jobIds, 'approved').then((response) => {
-        if (response.ok) {
-          enqueueSnackbar(`${title}'s approve request is accepted.`, { variant: 'success' });
-        } else {
-          enqueueSnackbar(`${title} are failed to approve.`, { variant: 'error' });
-        }
-      });
-    });
-  }, [confirm, enqueueSnackbar, batchUpdateStatus]);
-
-  const onBatchPause = useCallback((event: any, jobs: any[]) => {
-    const title = `${jobs.length} jobs`;
-    return confirm(`Pause ${title}?`).then((answer) => {
-      if (answer === false) return;
-
-      const jobIds = jobs.map((job) => job['jobId'])
-
-      enqueueSnackbar(`${title} are being paused.`);
-      return batchUpdateStatus(jobIds, 'pausing').then((response) => {
-        if (response.ok) {
-          enqueueSnackbar(`${title}'s pause request is accepted.`, { variant: 'success' });
-        } else {
-          enqueueSnackbar(`${title} are failed to pause.`, { variant: 'error' });
-        }
-      });
-    });
-  }, [confirm, enqueueSnackbar, batchUpdateStatus]);
-
-  const onBatchResume = useCallback((event: any, jobs: any[]) => {
-    const title = `${jobs.length} jobs`;
-    return confirm(`Resume ${title}?`).then((answer) => {
-      if (answer === false) return;
-
-      const jobIds = jobs.map((job) => job['jobId'])
-
-      enqueueSnackbar(`${title} are being resumed.`);
-      return batchUpdateStatus(jobIds, 'queued').then((response) => {
-        if (response.ok) {
-          enqueueSnackbar(`${title}'s resume request is accepted.`, { variant: 'success' });
-        } else {
-          enqueueSnackbar(`${title} are failed to resume.`, { variant: 'error' });
-        }
-      });
-    });
-  }, [confirm, enqueueSnackbar, batchUpdateStatus]);
-
-  const onBatchKill = useCallback((event: any, jobs: any[]) => {
-    const title = `${jobs.length} jobs`;
-    return confirm(`Kill ${title}?`).then((answer) => {
-      if (answer === false) return;
-
-      const jobIds = jobs.map((job) => job['jobId'])
-
-      enqueueSnackbar(`${title} are being killed.`);
-      return batchUpdateStatus(jobIds, 'killing').then((response) => {
-        if (response.ok) {
-          enqueueSnackbar(`${title}'s kill request is accepted.`, { variant: 'success' });
-        } else {
-          enqueueSnackbar(`${title} are failed to kill.`, { variant: 'error' });
-        }
-      });
-    });
-  }, [confirm, enqueueSnackbar, batchUpdateStatus]);
-
   const support = useCallback(Object.assign((job: any): Action<any> => {
     return {
       icon: 'contact_support',
@@ -228,6 +141,7 @@ ${givenName} ${familyName}
       onClick: onApprove
     }
   }, { position: 'row' }), [onApprove]);
+
   const pause = useCallback(Object.assign((job: any): Action<any> => {
     const hidden = PAUSABLE_STATUSES.indexOf(job['jobStatus']) === -1;
     return {
@@ -237,6 +151,7 @@ ${givenName} ${familyName}
       onClick: onPause
     }
   }, { position: 'row' }), [onPause]);
+
   const resume = useCallback(Object.assign((job: any): Action<any> => {
     const hidden = RESUMABLE_STATUSES.indexOf(job['jobStatus']) === -1;
     return {
@@ -246,6 +161,7 @@ ${givenName} ${familyName}
       onClick: onResume
     }
   }, { position: 'row' }), [onResume]);
+
   const kill = useCallback(Object.assign((job: any): Action<any> => {
     const hidden = KILLABLE_STATUSES.indexOf(job['jobStatus']) === -1;
     return {
@@ -256,48 +172,7 @@ ${givenName} ${familyName}
     }
   }, { position: 'row' }), [onKill]);
 
-  const batchApprove = useCallback(Object.assign((jobs: any[]): Action<any> => {
-    const hidden = !Array.isArray(jobs) || jobs.some(job => APPROVABLE_STATUSES.indexOf(job['jobStatus']) === -1);
-    return {
-      hidden,
-      icon: 'check',
-      tooltip: 'Approve',
-      onClick: onBatchApprove
-    }
-  }, { position: 'toolbarOnSelect' }), [onBatchApprove]);
-  const batchPause = useCallback(Object.assign((jobs: any[]): Action<any> => {
-    const hidden = !Array.isArray(jobs) || jobs.some(job => PAUSABLE_STATUSES.indexOf(job['jobStatus']) === -1);
-    return {
-      hidden,
-      icon: 'pause',
-      tooltip: 'Pause',
-      onClick: onBatchPause
-    }
-  }, { position: 'toolbarOnSelect' }), [onBatchPause]);
-  const batchResume = useCallback(Object.assign((jobs: any[]): Action<any> => {
-    const hidden = !Array.isArray(jobs) || jobs.some(job => RESUMABLE_STATUSES.indexOf(job['jobStatus']) === -1);
-    return {
-      hidden,
-      icon: 'play_arrow',
-      tooltip: 'Resume',
-      onClick: onBatchResume
-    }
-  }, { position: 'toolbarOnSelect' }), [onBatchResume]);
-  const batchKill = useCallback(Object.assign((jobs: any[]): Action<any> => {
-    const hidden = !Array.isArray(jobs) || jobs.some(job => KILLABLE_STATUSES.indexOf(job['jobStatus']) === -1);
-    return {
-      hidden,
-      icon: 'clear',
-      tooltip: 'Kill',
-      onClick: onBatchKill
-    }
-  }, { position: 'toolbarOnSelect' }), [onBatchKill]);
-
-  return {
-    support,
-    approve, pause, resume, kill,
-    batchApprove, batchPause, batchResume, batchKill
-  };
+  return { support, approve, pause, resume, kill };
 }
 
 export default useActions;
